requirements.txt
Flask==2.1.0
Pillow==8.3.1
tensorflow==2.6.0